requirements.txt
Appium-Python-Client==0.26
attrs==18.1.0
funcsigs==1.0.2
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
selenium==3.12.0
six==1.11.0
toolz==0.9.0
